requirements.txt
duckdb>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
pyyaml>=6.0
transformers>=4.40.0
torch>=2.0.0
umap-learn>=0.5.5
hdbscan>=0.8.33
bertopic>=0.16.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
networkx>=3.0
pyvista>=0.43.0
tqdm>=4.65.0

